backend/requirements-prod.txt
fastapi
uvicorn
python-dotenv
pillow
requests
pytesseract
python-multipart
discogs-client
beautifulsoup4
google-api-python-client
oauth2client
Werkzeug
numpy
scikit-image
packaging
six
retry
lxml
google-cloud-vision
spotipy
fuzzywuzzy
python-Levenshtein
discogs_client
google-generativeai>=0.3.0
aiohttp>=3.8.0
yt-dlp>=2023.7.6
